+++ src/main/java/com/siberteam/edu/zernest/calc/form/CalculatorForm.java
@@ -5,8 +5,8 @@ import org.apache.struts.validator.ValidatorForm;
 public class CalculatorForm extends ValidatorForm {
     private String x = "0";
     private String y = "0";
-    private String operation="+";
-    private int result;
+    private String operation = "+";
+    private Double result;
 
     public String getX() {
         return x;
@@ -32,11 +32,11 @@ public class CalculatorForm extends ValidatorForm {
         this.operation = operation;
     }
 
-    public int getResult() {
+    public Double getResult() {
         return result;
     }
 
-    public void setResult(int result) {
+    public void setResult(Double result) {
         this.result = result;
     }
 }
